Move Snapshot configuration methods into a separate file in shared helper folder
